Worker refactoring finished
